Settings: Fix color picker ComposeShader view
The shader was not showing the proper combined gradients.  This is a
limitation of the ComposeShader in Android 3+ and has to do with the
hardware acceleration as described here:

'ComposeShader can only contain shaders of differenttypes (a BitmapShader
and a LinearGradient for instance, but not two instances of BitmapShader)'.
But, 'If your application is affected by any of these missing features or
limitations, you can turn off hardware acceleration for just the affected
portion of your application by calling:
    setLayerType(View.LAYER_TYPE_SOFTWARE, null).'

Change-Id: I8d945e37ed97c409a0cc7e9ec5570019545f0f68